requirements.txt
absl-py==0.9.0
astor==0.8.1
bio==0.1.0
certifi==2020.4.5.1
cycler==0.10.0
gast==0.3.3
grpcio==1.28.1
h5py==2.10.0
joblib==0.14.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
numpy==1.18.3
pandas==1.0.3
protobuf==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
scikit-learn==0.22.2.post1
scipy==1.2.1
seaborn==0.11.0
six==1.14.0
sklearn==0.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0


